pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "piaudioplumber"
version = "0.1.0"
description = "PipeWire audio supervisor daemon for Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "paho-mqtt>=2.1.0",
    "aiosqlite>=0.20.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
pap-daemon = "pap.daemon.main:main"
papctl = "pap.cli.papctl:app"
pap-ha-adapter = "pap.plugins.ha_adapter.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pap"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
